+++ src/ui/backend.rs
@@ -222,9 +222,21 @@ where
         self.terminal.write_styled(&x)
     }
 
-    fn print_option_value<D: Display>(&mut self, option: &ListOption<D>) -> Result<()> {
+    fn print_option_value<D: Display>(
+        &mut self,
+        option: &ListOption<D>,
+        page: &Page<ListOption<D>>,
+    ) -> Result<()> {
+        let style_sheet = if page.content[page.selection].index == option.index
+            && self.render_config.selected_option.is_some()
+        {
+            self.render_config.selected_option.unwrap()
+        } else {
+            self.render_config.option
+        };
+
         self.terminal
-            .write_styled(&Styled::new(&option.value).with_style_sheet(self.render_config.option))
+            .write_styled(&Styled::new(&option.value).with_style_sheet(style_sheet))
     }
 
     fn print_option_index_prefix(&mut self, index: usize, max_index: usize) -> Option<Result<()>> {
@@ -449,8 +461,7 @@ where
             self.print_option_prefix(idx, &page)?;
 
             self.terminal.write(" ")?;
-
-            self.print_option_value(option)?;
+            self.print_option_value(option, &page)?;
 
             self.new_line()?;
         }
@@ -498,7 +509,7 @@ where
                 self.terminal.write(" ")?;
             }
 
-            self.print_option_value(option)?;
+            self.print_option_value(option, &page)?;
 
             self.new_line()?;
         }
@@ -541,7 +552,7 @@ where
 
             self.terminal.write(" ")?;
 
-            self.print_option_value(option)?;
+            self.print_option_value(option, &page)?;
 
             self.new_line()?;
         }