+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### Breaking Changes
 
-- Output dialogs on `stderr` instead of `stdout` [#89](https://github.com/mikaelmello/inquire/pull/89)
-  - `stdout` is usually used to carry (parseable) data, while other information (logs, dlailogs, etc.) got to `stderr`. This makes `inquire` compatible with such systems.
+- Selected option can now be styled independently of other options through `RenderConfig::with_selected_option()`.
+- Output dialogs on `stderr` instead of `stdout` [#89](https://github.com/mikaelmello/inquire/pull/89).
 
 ## [0.5.3] - 2023-01-09